+++ src/main/java/com/github/lipiridi/spotless/applier/action/ReformatFileAction.java
@@ -8,7 +8,6 @@ import com.intellij.openapi.actionSystem.CommonDataKeys;
 import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.project.Project;
 import com.intellij.psi.PsiFile;
-
 import java.util.Optional;
 import org.jetbrains.annotations.NotNull;